backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
openai==1.3.7
anthropic==0.7.7
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
lxml==4.9.3
cssselect==1.2.0
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
xlsxwriter==3.1.9
Pillow==10.1.0
python-dateutil==2.8.2
validators==0.22.0
fake-useragent==1.4.0
aioredis==2.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
spacy==3.7.2
wordcloud==1.9.2
plotly==5.17.0
email-validator==2.1.0
jinja2==3.1.2
sendgrid==6.12.5